Change interface names
